Add the option to add and edit new user

+++ src/main/java/com/example/loginplsql/controllers/UserController.java
@@ -4,6 +4,7 @@ import com.example.loginplsql.models.LoginResponse;
 import com.example.loginplsql.models.User;
 import com.example.loginplsql.services.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,7 +12,6 @@ import java.util.List;
 
 @RestController
 public class UserController {
-//TODO: Add Token
     private final UserServiceImpl userService;
 
     @Autowired
@@ -20,28 +20,46 @@ public class UserController {
     }
 
     @GetMapping("/users_list")
-    public List<User> getAllUsers() {
-        return userService.findAllUsers();
+    public ResponseEntity<List<User>> getAllUsers(@RequestHeader LoginResponse response) {
+        if (userService.verifyUserAndToken(response.getResponse())) {
+            return ResponseEntity.ok(userService.findAllUsers());
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
     }
 
     @PostMapping("/add-user")
-    public User addUser(@RequestBody User user) {
-        return userService.addUser(user);
+    public ResponseEntity<User> addUser(@RequestBody User user, @RequestHeader LoginResponse response) {
+        if (userService.verifyUserAndToken(response.getResponse())) {
+            return ResponseEntity.ok(userService.addUser(user));
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @GetMapping("/getUser/{id}")
-    public User getUserById(@PathVariable int id) {
-        return userService.findUserById(id);
+    public ResponseEntity<User> getUserById(@PathVariable int id, @RequestHeader LoginResponse response) {
+        if (userService.verifyUserAndToken(response.getResponse())) {
+            User user = userService.findUserById(id);
+            return ResponseEntity.ok(user);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @GetMapping("/getUser/{username}")
-    public User getUserByUsername(@PathVariable String username) {
-        return userService.findUserByUsername(username);
+    public ResponseEntity<User> getUserByUsername(@PathVariable String username, @RequestHeader LoginResponse response) {
+        if (userService.verifyUserAndToken(response.getResponse())) {
+            User user = userService.findUserByUsername(username);
+            return ResponseEntity.ok(user);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @DeleteMapping("/delete-user/{id}")
-    public void deleteUser(@PathVariable int id) {
-        userService.deleteUserById(id);
+    public ResponseEntity<Void> deleteUser(@PathVariable int id, @RequestHeader LoginResponse response) {
+        if (userService.verifyUserAndToken(response.getResponse())) {
+            userService.deleteUserById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @PostMapping("/login")
@@ -49,9 +67,11 @@ public class UserController {
         return userService.login(request);
     }
 
-    @PutMapping("/update-user/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable int id, @RequestBody User userDetails) {
-        return ResponseEntity.ok(userService.updateUser(id, userDetails));
+    @PostMapping("/update-user/{id}")
+    public ResponseEntity<User> updateUser(@PathVariable int id, @RequestBody User userDetails, @RequestHeader LoginResponse response) {
+        if (userService.verifyUserAndToken(response.getResponse())) {
+            return ResponseEntity.ok(userService.updateUser(id, userDetails));
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
-
 }
